requirements.txt
flask
google-genai
openai 
boto3>=1.28.0
ffmpeg-python
flask-login
flask-sqlalchemy
werkzeug
google-generativeai
python-dotenv
stripe
requests
Flask-Session
flask-pymongo
pymongo
fastapi
uvicorn
celery
redis
python-multipart
sendgrid
flask-mail
gunicorn
rq
pydub
imageio[ffmpeg]
